Modpack Uploaded - Version 1.0
This release includes the standard, non-43 Civ installation of Vox Populi as a DLC. It also includes Historical Religions Edit and Strange Religions (albeit with all but four religions removed). Credit to everyone who contributed to these mods, as well as the people behind the DLC hack that allows mods to be used in multiplayer (and also install them to your base Civ V installation).